--- platform-services/analysis-service/.gitignore
@@ -1,101 +1,0 @@
-#  before PyInstaller builds the exe, so as to inject date/other infos into it.
-*.manifest
-*.spec
-
-# Data
-data/*
-Results/*
-
-# Python Stuff
-*.pyc
-__pycache__
-
-#Stuff
-.DS_Store
-
-# Installer logs
-pip-log.txt
-pip-delete-this-directory.txt
-
-# Unit test / coverage reports
-htmlcov/
-.tox/
-.coverage
-.coverage.*
-.cache
-nosetests.xml
-coverage.xml
-*.cover
-.hypothesis/
-.pytest_cache/
-
-# Translations
-*.mo
-*.pot
-
-# Django stuff:
-*.log
-local_settings.py
-db.sqlite3
-
-# Flask stuff:
-instance/
-.webassets-cache
-
-# Scrapy stuff:
-.scrapy
-
-# Sphinx documentation
-docs/_build/
-
-# PyBuilder
-target/
-
-# Jupyter Notebook
-.ipynb_checkpoints
-
-# pyenv
-.python-version
-
-# celery beat schedule file
-celerybeat-schedule
-
-# SageMath parsed files
-*.sage.py
-
-# Environments
-.venv
-env
-.env
-venv
-ENV
-env.bak
-venv.bak
-
-# Spyder project settings
-.spyderproject
-.spyproject
-# Rope project settings
-.ropeproject
-
-# mkdocs documentation
-/site
-
-# mypy
-.mypy_cache/
-
-# System
-.DS_Store
-.idea
-*.pyc
-
-# Dockerfiles
-deps.zip
-
-.idea/
-yarn.lock
-.serverless
-.requirements.zip
-node_modules
-
-
