--- platform-services/analysis-service/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "spacemaker-data-api",
-  "version": "1.0.0",
-  "license": "MIT",
-  "scripts": {
-    "install:deps": "pip install -r requirements.txt",
-    "build": "serverless package",
-    "deploy": "serverless deploy --stage=prod",
-    "deploy:get-data-from-solutions": "serverless deploy --stage=prod function -f analyse-solution-in-course",
-    "logs": "serverless logs -t -f analyse-solution-in-course"
-  },
-  "dependencies": {
-    "serverless": "^1.32.0",
-    "serverless-python-requirements": "^4.1.1"
-  }
-}
