--- platform-services/analysis-service/requirements.txt
@@ -1,4 +1,0 @@
-spacemakerlog3
-rasterio
-shapely
-
