--- platform-services/analysis-service/serverless.yml
@@ -1,45 +1,0 @@
-service: docker-kubernet-analysis
-
-provider:
-  name: aws
-  region: eu-west-1
-  stage: prod
-  versionFunctions: false
-  runtime: python3.6
-  deploymentBucket:
-    name: code-upload
-
-package:
-  exclude:
-  - node_modules/**
-  - .env/**
-  - .idea/**
-  - "**/__pycache__/**"
-  - .pytest_cache/**
-  - yarn.lock
-  - rest.py
-plugins:
-- serverless-python-requirements
-
-
-custom:
-  pythonRequirements:
-    dockerizePip: non-linux
-    slim: true
-    zip: true
-    fileName: requirements.txt
-
-functions:
-  analyse-solution-in-course:
-    handler: analysis_handler.handler
-    role: arn:aws:iam::741408923162:role/lambda_invoke_lambda
-    timeout: 30
-    description: Calculate score for proposal
-    events:
-    - http:
-        path: /analyse_solution_in_course
-        method: post
-        request:
-          parameters:
-            headers:
-              Authorization: false
